refactor(canopy): move tamanu tags to canopy tags
The tags command fetches device tags from canopy, so it belongs under
the canopy subcommand. Move the implementation there and keep
`bestool tamanu tags` as a re-export alias.
